show chat message instead of notif
