+++ plugins/amazonq/codetransform/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codemodernizer/CodeModernizerManager.kt
@@ -70,6 +70,7 @@ import software.aws.toolkits.jetbrains.services.codemodernizer.utils.getSupporte
 import software.aws.toolkits.jetbrains.services.codemodernizer.utils.isCodeTransformAvailable
 import software.aws.toolkits.jetbrains.services.codemodernizer.utils.isGradleProject
 import software.aws.toolkits.jetbrains.services.codemodernizer.utils.openTroubleshootingGuideNotificationAction
+import software.aws.toolkits.jetbrains.services.codemodernizer.utils.parseBuildFile
 import software.aws.toolkits.jetbrains.services.codemodernizer.utils.parseXmlDependenciesReport
 import software.aws.toolkits.jetbrains.services.codemodernizer.utils.setDependencyVersionInPom
 import software.aws.toolkits.jetbrains.services.codemodernizer.utils.tryGetJdk
@@ -393,6 +394,10 @@ class CodeModernizerManager(private val project: Project) : PersistentStateCompo
             isMvnRunning.set(false)
             handleLocalMavenBuildResult(result)
         }
+    }
+
+    fun parseBuildFile(): String? {
+        return parseBuildFile(codeTransformationSession?.sessionContext?.configurationFile)
     }
 
     internal suspend fun initModernizationJob(session: CodeModernizerSession, copyResult: MavenCopyCommandsResult): CodeModernizerJobCompletedResult =
