+++ plugins/amazonq/codetransform/jetbrains-community/tst/software/aws/toolkits/jetbrains/services/codemodernizer/CodeWhispererCodeModernizerUtilsTest.kt
@@ -8,6 +8,7 @@ import io.mockk.mockkStatic
 import io.mockk.runs
 import kotlinx.coroutines.runBlocking
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
 import org.mockito.Mockito
@@ -20,9 +21,12 @@ import software.amazon.awssdk.services.codewhispererruntime.model.Transformation
 import software.amazon.awssdk.services.codewhispererruntime.model.TransformationStatus
 import software.amazon.awssdk.services.ssooidc.model.InvalidGrantException
 import software.aws.toolkits.jetbrains.services.codemodernizer.utils.getTableMapping
+import software.aws.toolkits.jetbrains.services.codemodernizer.utils.parseBuildFile
 import software.aws.toolkits.jetbrains.services.codemodernizer.utils.pollTransformationStatusAndPlan
 import software.aws.toolkits.jetbrains.services.codemodernizer.utils.refreshToken
+import software.aws.toolkits.jetbrains.utils.rules.addFileToModule
 import java.util.concurrent.atomic.AtomicBoolean
+import kotlin.test.assertNotNull
 
 class CodeWhispererCodeModernizerUtilsTest : CodeWhispererCodeModernizerTestBase() {
     @Before
@@ -203,5 +207,13 @@ class CodeWhispererCodeModernizerUtilsTest : CodeWhispererCodeModernizerTestBase
         val actual = getTableMapping(listOf(step0Update0, step0Update1, step0Update2, step0Update3))
         val expected = mapOf("0" to jobStats, "1" to depChanges, "2" to apiChanges, "-1" to fileChanges)
         assertThat(expected).isEqualTo(actual)
+    }
+
+    @Test
+    fun `parseBuildFile can detect absolute paths in build file`() {
+        val module = projectRule.module
+        val fileText = "<project><properties><path>system/name/here</path></properties></project>"
+        val file = projectRule.fixture.addFileToModule(module, "pom.xml", fileText)
+        assertNotNull(parseBuildFile(file.virtualFile))
     }
 }
